Solve issues #1, #2
Renames files to be lowercase.
Do not export default.
Instead of softReject use a response third parameter.
