+++ src/race-until.d.ts
@@ -1,0 +1,2 @@
+export function timebomb(timeout?: number, softReject?: boolean): any;
+export function raceUntil(p: any, timeout?: number, softReject?: boolean): any;
