--- src/raceUntil.d.ts
@@ -1,2 +1,0 @@
-export declare function timebomb(timeout?: number, softReject?: boolean): any;
-export default function raceUntil(p: any, timeout?: number, softReject?: boolean): any;
